Allow configuring the bind/render hooks and media type per route
Being able to configure the bind and render hooks and the media types
per route is convenient in order to be able to easily manage API which
work with various types of data.

I introduced two variables for the media types (request and response)
in order to be able to make the distinction in Fizz between the
requests and responses media types.